style: Add Logo compo and Link for index #68
