--- src/components/organism/SideOption.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-interface Props {
-  
-}
-
-const SideOption = (props: Props) => {
-  return (
-    <div>
-      <p>side option</p>
-    </div>
-  )
-}
-
-export default SideOption
