+++ src/components/templates/SideBar.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import Logo from '../molecles/Logo'
+
+interface Props {
+  
+}
+
+const SideBar = (props: Props) => {
+  return (
+    <div>
+      <Logo/>
+    </div>
+  )
+}
+
+export default SideBar
