+++ src/components/templates/SideWidget.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import { PopularPosts } from '../organism/PopularPosts'
 import { PopularUsers } from '../organism/PopularUsers'
 import { TagBox } from '../organism/TagBox'
